+++ src/main/generate_butterfly_files.py
@@ -3,77 +3,6 @@ import os
 import pathlib
 import shutil
 
-# The list of butterfly species included in the mod.
-# NOTE: Avoid using underscores in these names.
-BUTTERFLIES = [
-    'admiral',
-    'buckeye',
-    'cabbage',
-    'chalkhill',
-    'clipper',
-    'clipperwhite',
-    'clipperlightgray',
-    'clippergray',
-    'clipperblack',
-    'clipperbrown',
-    'clipperred',
-    'clipperorange',
-    'clipperyellow',
-    'clipperlime',
-    'clippergreen',
-    'clippercyan',
-    'clipperblue',
-    'clipperpurple',
-    'clippermagenta',
-    'clipperpink',
-    'common',
-    'emperor',
-    'forester',
-    'glasswing',
-    'hairstreak',
-    'heath',
-    'longwing',
-    'monarch',
-    'morpho',
-    'rainbow',
-    'swallowtail',
-    'peacock',
-    'commongrassyellow',
-    'commander',
-    'commonbirdwing',
-    'bluemoon'
-]
-
-MALE_BUTTERFLIES = [
-    'bluemoonmale'
-]
-
-# The list of moth species included in the mod.
-# NOTE: Avoid using underscores in these names.
-MOTHS = [
-    'clothes',
-    'luna',
-    'domestic_silk',
-    'peppered',
-    'indianmeal',
-    'spongy',
-    'atlas',
-    'carpet',
-    'codling',
-    'diamondback'
-]
-
-MALE_MOTHS = [
-    'spongymale'
-]
-
-# The list of special butterflies included in the mod.
-# NOTE: Avoid using underscores in these names.
-SPECIAL = [
-    'ice',
-    'lava'
-]
-
 # List of files to generate json files for.
 FLOWERS = [
     'allium',
@@ -92,18 +21,38 @@ FLOWERS = [
 ]
 
 # File locations
+ACHIEVEMENTS = "resources/data/butterflies/advancements/butterfly/"
+BIOME_MODIFIERS = "resources/data/butterflies/forge/biome_modifier/"
+BUTTERFLY_DATA = "resources/data/butterflies/butterfly_data/"
+CODE_GENERATION = "java/com/bokmcdok/butterflies/world/ButterflySpeciesList.java"
 FROG_FOOD = "resources/data/minecraft/tags/entity_types/frog_food.json"
 LOCALISATION = "resources/assets/butterflies/lang/en_us.json"
-ACHIEVEMENTS = "resources/data/butterflies/advancements/butterfly/"
-BUTTERFLY_ACHIEVEMENT_TEMPLATES = "resources/data/butterflies/advancement_templates/butterfly/"
-MALE_BUTTERFLY_ACHIEVEMENT_TEMPLATES = "resources/data/butterflies/advancement_templates/butterfly_male/"
-MOTH_ACHIEVEMENT_TEMPLATES = "resources/data/butterflies/advancement_templates/moth/"
-MALE_MOTH_ACHIEVEMENT_TEMPLATES = "resources/data/butterflies/advancement_templates/moth_male/"
-BOTH_ACHIEVEMENT_TEMPLATES = "resources/data/butterflies/advancement_templates/both/"
-CODE_GENERATION = "java/com/bokmcdok/butterflies/world/ButterflySpeciesList.java"
+BUTTERFLY_ACHIEVEMENT_TEMPLATES = "resources/data/butterflies/templates/advancements/butterfly/"
+MALE_BUTTERFLY_ACHIEVEMENT_TEMPLATES = "resources/data/butterflies/templates/advancements/butterfly_male/"
+MOTH_ACHIEVEMENT_TEMPLATES = "resources/data/butterflies/templates/advancements/moth/"
+MALE_MOTH_ACHIEVEMENT_TEMPLATES = "resources/data/butterflies/templates/advancements/moth_male/"
+BOTH_ACHIEVEMENT_TEMPLATES = "resources/data/butterflies/templates/advancements/both/"
+BIOME_MODIFIER_TEMPLATES = "resources/data/butterflies/templates/biome_modifiers/"
+
+BUTTERFLIES_FOLDER = "butterflies/"
+MALE_BUTTERFLIES_FOLDER = "butterflies/male/"
+MOTHS_FOLDER = "moths/"
+MALE_MOTHS_FOLDER = "moths/male/"
+SPECIAL_FOLDER = "special/"
+
 
 # Initial index
 BUTTERFLY_INDEX = 0
+
+def generate_butterfly_list(folder):
+    print(f"Generating species list for folder [{BUTTERFLY_DATA + folder}]")
+    result = []
+    for (path, _, filenames) in os.walk(BUTTERFLY_DATA + folder):
+        result = [f.replace(".json", "") for f in filenames if f.endswith(".json")]
+        break
+
+    print(f"Result: {repr(result)}")
+    return result
 
 
 # Generates data files from the input array based on the first entry in the#
@@ -118,8 +67,9 @@ def generate_data_files(entries):
         # We only want json
         filenames = [f for f in filenames if f.endswith(".json") and entries[0] in f]
 
-        for name in filenames:
-            files.append(pathlib.Path(path, name))
+        if "loot_tables" not in path:
+            for name in filenames:
+                files.append(pathlib.Path(path, name))
 
     # Loop Start
     for entry in entries:
@@ -183,25 +133,10 @@ class FrogFood(object):
 
 
 # Generate list of entities to add to frog food.
-def generate_frog_food():
+def generate_frog_food(species):
     print("Generating frog food...")
 
-    values = []
-    for i in BUTTERFLIES:
-        values.append("butterflies:" + i)
-
-    for i in MALE_BUTTERFLIES:
-        values.append("butterflies:" + i)
-
-    for i in MOTHS:
-        values.append("butterflies:" + i)
-
-    for i in MALE_MOTHS:
-        values.append("butterflies:" + i)
-
-    for i in SPECIAL:
-        values.append("butterflies:" + i)
-
+    values = ["butterflies:" + i for i in species]
     frog_food = FrogFood(values)
 
     with open(FROG_FOOD, 'w') as file:
@@ -216,13 +151,13 @@ def try_add_localisation_string(json_data, key, value):
 
 
 # Generates localisation strings if they don't already exist.
-def generate_localisation_strings():
+def generate_localisation_strings(all_butterflies, all_moths):
     print("Generating localisation strings...")
 
     with open(LOCALISATION, 'r', encoding="utf8") as input_file:
         json_data = json.load(input_file)
 
-    for i in BUTTERFLIES + MALE_BUTTERFLIES + SPECIAL:
+    for i in all_butterflies:
         name = i.replace('_', ' ')
         name = name.title()
         try_add_localisation_string(json_data, "entity.butterflies." + i, name + " Butterfly")
@@ -232,7 +167,7 @@ def generate_localisation_strings():
         try_add_localisation_string(json_data, "item.butterflies." + i + "_egg", name + " Butterfly Egg")
         try_add_localisation_string(json_data, "item.butterflies." + i + "_caterpillar", name + " Caterpillar")
 
-    for i in MOTHS + MALE_MOTHS:
+    for i in all_moths:
         name = i.replace('_', ' ')
         name = name.title()
         try_add_localisation_string(json_data, "entity.butterflies." + i + "_caterpillar", name + " Larva")
@@ -240,7 +175,7 @@ def generate_localisation_strings():
         try_add_localisation_string(json_data, "item.butterflies." + i + "_egg", name + " Moth Egg")
         try_add_localisation_string(json_data, "item.butterflies." + i + "_caterpillar", name + " Larva")
 
-    for i in BUTTERFLIES + MALE_BUTTERFLIES + MOTHS + MALE_MOTHS + SPECIAL:
+    for i in all_butterflies + all_moths:
         try_add_localisation_string(json_data, "gui.butterflies.fact." + i, "")
 
     with open(LOCALISATION, 'w', encoding="utf8") as file:
@@ -300,7 +235,7 @@ def generate_advancements(entities, templates):
 # then used to create all the entities and items related to each butterfly,
 # saving a ton of new code needing to be written every time we add a new
 # butterfly or moth.
-def generate_code():
+def generate_code(all):
     print("Generating code...")
 
     with open(CODE_GENERATION, 'w', encoding="utf8") as output_file:
@@ -313,24 +248,8 @@ public class ButterflySpeciesList {
     public static final String[] SPECIES = {
 """)
 
-        for butterfly in BUTTERFLIES:
+        for butterfly in all:
             output_file.write("""            \"""" + butterfly + """\",
-""")
-
-        for butterfly in MALE_BUTTERFLIES:
-            output_file.write("""            \"""" + butterfly + """\",
-""")
-
-        for moth in MOTHS:
-            output_file.write("""            \"""" + moth + """\",
-""")
-
-        for moth in MALE_MOTHS:
-            output_file.write("""            \"""" + moth + """\",
-""")
-
-        for special in SPECIAL:
-            output_file.write("""            \"""" + special + """\",
 """)
 
         output_file.write("""    };
@@ -366,20 +285,179 @@ def generate_textures(entries, base):
                     shutil.copy(file, new_file)
 
 
+# Reset the biome modifiers ready for generation.
+def reset_biome_modifiers():
+    for (_, _, files) in os.walk(BIOME_MODIFIER_TEMPLATES):
+        for file_ in files:
+            src_file = os.path.join(BIOME_MODIFIER_TEMPLATES, file_)
+            dst_file = os.path.join(BIOME_MODIFIERS, file_)
+            if os.path.exists(dst_file):
+                # in case of the src and dst are the same file
+                if os.path.samefile(src_file, dst_file):
+                    continue
+                os.remove(dst_file)
+            shutil.copy(src_file, BIOME_MODIFIERS)
+
+
+# Generate biome modifiers based on rarity and habitat
+def generate_biome_modifiers(list, folder, is_male):
+    print("Generating biome modifiers...")
+
+    # Iterate over butterflies
+    for i in list:
+        addSpawns(folder, i, is_male)
+
+
+# Add a group of spawns for a single butterfly.
+def addSpawns(folder, species, is_male):
+
+    # Open Butterfly data
+    with open(BUTTERFLY_DATA + folder + species + ".json", 'r', encoding="utf8") as input_file:
+        json_data = json.load(input_file)
+
+    rarity = json_data["rarity"]
+    habitat = json_data["habitat"]
+
+    # Generate weights/min/max
+    if (rarity == "common"):
+        weight = 4
+        maximum = 4
+
+    elif (rarity == "uncommon"):
+        weight = 2
+        maximum = 3
+
+    else:
+        weight = 1
+        maximum = 2
+
+    # Open relevant files and add butterfly spawns
+    if "forests" in habitat:
+        addSingleSpawn("cherry_grove", species, weight, maximum, is_male)
+        addSingleSpawn("dense", species, weight, maximum, is_male)
+        addSingleSpawn("forest", species, weight, maximum, is_male)
+        addSingleSpawn("lush", species, weight, maximum, is_male)
+        addSingleSpawn("taiga", species, weight, maximum, is_male)
+
+    if "hills" in habitat:
+        addSingleSpawn("hill", species, weight, maximum, is_male)
+
+    if "ice" in habitat:
+        addSingleSpawn("snowy", species, weight, maximum, is_male)
+        addSingleSpawn("taiga", species, weight, maximum, is_male)
+
+    if "jungles" in habitat:
+        addSingleSpawn("jungle", species, weight, maximum, is_male)
+
+    if "nether" in habitat:
+        addSingleSpawn("nether", species, weight, maximum, True)
+
+    if "plains" in habitat:
+        addSingleSpawn("plains", species, weight, maximum, is_male)
+
+    if "plateaus" in habitat:
+        addSingleSpawn("lush", species, weight, maximum, is_male)
+        addSingleSpawn("plateau", species, weight, maximum, is_male)
+
+    if "savanna" in habitat:
+        addSingleSpawn("savanna", species, weight, maximum, is_male)
+
+    if "villages" in habitat:
+        addSingleSpawn("village_desert", species, weight, maximum, is_male)
+        addSingleSpawn("village_plains", species, weight, maximum, is_male)
+        addSingleSpawn("village_savanna", species, weight, maximum, is_male)
+        addSingleSpawn("village_snowy", species, weight, maximum, is_male)
+        addSingleSpawn("village_taiga", species, weight, maximum, is_male)
+
+    if "wetlands" in habitat:
+        addSingleSpawn("mushroom", species, weight, maximum, is_male)
+        addSingleSpawn("river", species, weight, maximum, is_male)
+        addSingleSpawn("swamp", species, weight, maximum, is_male)
+
+
+# Add a single spawn for a butterfly.
+def addSingleSpawn(tag, species, weight, maximum, is_male):
+    target_file = BIOME_MODIFIERS + tag + "_butterflies.json"
+
+    with open(target_file, 'r', encoding="utf8") as input_file:
+        json_data = json.load(input_file)
+
+    json_data["spawners"].append({
+        "type" : "butterflies:" + species,
+        "weight" : weight,
+        "minCount" : 1,
+        "maxCount" : maximum
+    })
+
+    if not is_male:
+        json_data["spawners"].append({
+            "type" : "butterflies:" + species + "_caterpillar",
+            "weight" : weight,
+            "minCount" : 1,
+            "maxCount" : maximum
+        })
+
+        json_data["spawners"].append({
+            "type" : "butterflies:" + species + "_chrysalis",
+            "weight" : weight,
+            "minCount" : 1,
+            "maxCount" : maximum
+        })
+
+        json_data["spawners"].append({
+            "type" : "butterflies:" + species + "_egg",
+            "weight" : weight,
+            "minCount" : 1,
+            "maxCount" : maximum
+        })
+
+    with open(target_file, 'w', encoding="utf8") as output_file:
+        output_file.write(json.dumps(json_data,
+                                     default=lambda o: o.__dict__,
+                                     sort_keys=True,
+                                     indent=2))
+
+
 # Python's main entry point
 if __name__ == "__main__":
-    generate_data_files(BUTTERFLIES)
-    generate_data_files(MALE_BUTTERFLIES)
-    generate_data_files(MOTHS)
-    generate_data_files(MALE_MOTHS)
-    generate_data_files(SPECIAL)
+    # Get the species lists
+    butterflies = generate_butterfly_list(BUTTERFLIES_FOLDER)
+    male_butterflies = generate_butterfly_list(MALE_BUTTERFLIES_FOLDER)
+    moths = generate_butterfly_list(MOTHS_FOLDER)
+    male_moths = generate_butterfly_list(MALE_MOTHS_FOLDER)
+    special = generate_butterfly_list(SPECIAL_FOLDER)
+
+    all = butterflies + male_butterflies + moths + male_moths + special
+    all_butterflies = butterflies + male_butterflies
+    all_moths = moths + male_moths
+
+    # Generate the data files
+    generate_data_files(butterflies)
+    generate_data_files(male_butterflies)
+    generate_data_files(moths)
+    generate_data_files(male_moths)
+    generate_data_files(special)
+
     #generate_textures(BUTTERFLIES, "clipper") # Change this to use a different base for textures
     # generate_data_files(FLOWERS) # Disabled for now due to tulip problem
-    generate_frog_food()
-    generate_localisation_strings()
-    generate_advancements(BUTTERFLIES, BUTTERFLY_ACHIEVEMENT_TEMPLATES)
-    generate_advancements(BUTTERFLIES + MALE_BUTTERFLIES, MALE_BUTTERFLY_ACHIEVEMENT_TEMPLATES)
-    generate_advancements(MOTHS, MOTH_ACHIEVEMENT_TEMPLATES)
-    generate_advancements(MOTHS + MALE_MOTHS, MALE_MOTH_ACHIEVEMENT_TEMPLATES)
-    generate_advancements(BUTTERFLIES + MOTHS, BOTH_ACHIEVEMENT_TEMPLATES)
-    generate_code()
+
+    generate_frog_food(all)
+    generate_localisation_strings(all_butterflies + special, all_moths)
+
+    # Generate the advancements
+    generate_advancements(butterflies, BUTTERFLY_ACHIEVEMENT_TEMPLATES)
+    generate_advancements(all_butterflies, MALE_BUTTERFLY_ACHIEVEMENT_TEMPLATES)
+    generate_advancements(moths, MOTH_ACHIEVEMENT_TEMPLATES)
+    generate_advancements(all_moths, MALE_MOTH_ACHIEVEMENT_TEMPLATES)
+    generate_advancements(butterflies + moths, BOTH_ACHIEVEMENT_TEMPLATES)
+
+    generate_code(all)
+
+    # Generate biome modifiers.
+    reset_biome_modifiers()
+    generate_biome_modifiers(butterflies, BUTTERFLIES_FOLDER, False)
+    generate_biome_modifiers(male_butterflies, MALE_BUTTERFLIES_FOLDER, True)
+    generate_biome_modifiers(moths, MOTHS_FOLDER, False)
+    generate_biome_modifiers(male_moths, MALE_MOTHS_FOLDER, True)
+    generate_biome_modifiers(special, SPECIAL_FOLDER, False)
+#