+++ src/main/java/com/bokmcdok/butterflies/ButterfliesMod.java
@@ -1,64 +1,79 @@
 package com.bokmcdok.butterflies;
 
+import com.bokmcdok.butterflies.client.event.ClientEventListener;
 import com.bokmcdok.butterflies.config.ButterfliesConfig;
-import com.bokmcdok.butterflies.registries.BlockEntityTypeRegistry;
-import com.bokmcdok.butterflies.registries.BlockRegistry;
-import com.bokmcdok.butterflies.registries.EntityTypeRegistry;
-import com.bokmcdok.butterflies.registries.ItemRegistry;
-import com.bokmcdok.butterflies.registries.LootModifierRegistry;
-import com.bokmcdok.butterflies.world.ButterflySpeciesList;
-import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.item.Items;
-import net.minecraft.world.item.alchemy.PotionUtils;
-import net.minecraft.world.item.alchemy.Potions;
-import net.minecraft.world.item.crafting.Ingredient;
-import net.minecraftforge.common.brewing.BrewingRecipeRegistry;
+import com.bokmcdok.butterflies.event.ForgeEventListener;
+import com.bokmcdok.butterflies.event.ModEventListener;
+import com.bokmcdok.butterflies.event.entity.EntityEventListener;
+import com.bokmcdok.butterflies.event.entity.player.PlayerEventListener;
+import com.bokmcdok.butterflies.event.server.ServerEventListener;
+import com.bokmcdok.butterflies.event.lifecycle.LifecycleEventListener;
+import com.bokmcdok.butterflies.event.network.NetworkEventListener;
+import com.bokmcdok.butterflies.event.village.VillageEventListener;
+import com.bokmcdok.butterflies.registries.*;
+import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.eventbus.api.IEventBus;
 import net.minecraftforge.fml.ModLoadingContext;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.config.ModConfig;
-import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent;
 import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
 
-import java.util.Arrays;
-
-// The value here should match an entry in the META-INF/mods.toml file
-@Mod(ButterfliesMod.MODID)
+/**
+ * The main entry point for the mod.
+ */
+@Mod(ButterfliesMod.MOD_ID)
 public class ButterfliesMod
 {
     // Define mod id in a common place for everything to reference
-    public static final String MODID = "butterflies";
+    public static final String MOD_ID = "butterflies";
 
     /**
-     * Constructor - The main entry point for the mod.
+     * Constructor.
      */
     public ButterfliesMod() {
         final IEventBus modEventBus = FMLJavaModLoadingContext.get().getModEventBus();
+        final IEventBus forgeEventBus = MinecraftForge.EVENT_BUS;
 
-        // Register ourselves for server and other game events we are interested in
-        modEventBus.register(this);
-        modEventBus.addListener(this::commonSetup);
+        // Create the registries.
+        BannerPatternRegistry bannerPatternRegistry = new BannerPatternRegistry(modEventBus);
+        BlockEntityTypeRegistry blockEntityTypeRegistry = new BlockEntityTypeRegistry(modEventBus);
+        BlockRegistry blockRegistry = new BlockRegistry(modEventBus);
+        DecoratedPotPatternsRegistry decoratedPotPatternsRegistry = new DecoratedPotPatternsRegistry(modEventBus);
+        EntityTypeRegistry entityTypeRegistry = new EntityTypeRegistry(modEventBus);
+        ItemRegistry itemRegistry = new ItemRegistry(modEventBus);
+        LootModifierRegistry lootModifierRegistry = new LootModifierRegistry(modEventBus);
+        MenuTypeRegistry menuTypeRegistry = new MenuTypeRegistry(modEventBus);
+        PoiTypeRegistry poiTypesRegistry = new PoiTypeRegistry(modEventBus);
+        VillagerProfessionRegistry villagerProfessionRegistry = new VillagerProfessionRegistry(modEventBus);
 
-        // Deferred registries.
-        BlockRegistry.INSTANCE.register(modEventBus);
-        BlockEntityTypeRegistry.INSTANCE.register(modEventBus);
-        EntityTypeRegistry.INSTANCE.register(modEventBus);
-        ItemRegistry.INSTANCE.register(modEventBus);
-        LootModifierRegistry.INSTANCE.register(modEventBus);
+        // Initialise the registries. Do this here because (e.g.)
+        // blockEntityTypeRegistry requires blockRegistry to be created and
+        // vice-versa.
+        bannerPatternRegistry.initialise();
+        blockEntityTypeRegistry.initialise(blockRegistry, menuTypeRegistry);
+        blockRegistry.initialise(blockEntityTypeRegistry, menuTypeRegistry);
+        decoratedPotPatternsRegistry.initialise();
+        entityTypeRegistry.initialise(blockRegistry);
+        itemRegistry.initialise(bannerPatternRegistry, blockRegistry, entityTypeRegistry);
+        lootModifierRegistry.initialise(itemRegistry);
+        menuTypeRegistry.initialise();
+        poiTypesRegistry.initialise(blockRegistry);
+        villagerProfessionRegistry.initialise(poiTypesRegistry);
+
+        // Create the Mod event listeners
+        new ClientEventListener(modEventBus, blockEntityTypeRegistry, entityTypeRegistry);
+        new LifecycleEventListener(modEventBus, decoratedPotPatternsRegistry, itemRegistry,menuTypeRegistry);
+        new ModEventListener(modEventBus, itemRegistry);
+
+        // Create the Forge event listeners.
+        new EntityEventListener(forgeEventBus, modEventBus, entityTypeRegistry);
+        new ForgeEventListener(forgeEventBus);
+        new NetworkEventListener(forgeEventBus);
+        new PlayerEventListener(forgeEventBus);
+        new ServerEventListener(forgeEventBus);
+        new VillageEventListener(forgeEventBus, itemRegistry, villagerProfessionRegistry);
 
         // Mod Config Settings
         ModLoadingContext.get().registerConfig(ModConfig.Type.SERVER, ButterfliesConfig.SERVER_CONFIG);
-    }
-
-    /**
-     * Common setup event where we register brewing recipes.
-     * @param event The event class.
-     */
-    private void commonSetup(FMLCommonSetupEvent event) {
-        int monarchIndex = Arrays.asList(ButterflySpeciesList.SPECIES).indexOf("monarch");
-        BrewingRecipeRegistry.addRecipe(
-                Ingredient.of(PotionUtils.setPotion(new ItemStack(Items.POTION), Potions.AWKWARD)),
-                Ingredient.of(ItemRegistry.CATERPILLAR_ITEMS.get(monarchIndex).get()),
-                PotionUtils.setPotion(new ItemStack(Items.POTION), Potions.POISON));
     }
 }