+++ src/main/java/com/bokmcdok/butterflies/event/entity/player/PlayerEventListener.java
@@ -1,11 +1,8 @@
 package com.bokmcdok.butterflies.event.entity.player;
 
-import com.bokmcdok.butterflies.ButterfliesMod;
 import com.bokmcdok.butterflies.world.ButterflyData;
 import com.bokmcdok.butterflies.world.CompoundTagId;
-import com.bokmcdok.butterflies.world.item.BottledButterflyItem;
 import com.bokmcdok.butterflies.world.item.ButterflyBookItem;
-import com.bokmcdok.butterflies.world.item.ButterflyContainerItem;
 import com.bokmcdok.butterflies.world.item.ButterflyScrollItem;
 import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.nbt.CompoundTag;
@@ -21,21 +18,52 @@ import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.NameTagItem;
 import net.minecraftforge.event.entity.player.PlayerEvent;
 import net.minecraftforge.event.entity.player.PlayerInteractEvent;
-import net.minecraftforge.eventbus.api.SubscribeEvent;
-import net.minecraftforge.fml.common.Mod;
+import net.minecraftforge.eventbus.api.IEventBus;
 
 /**
  * Holds event listeners for player actions.
  */
-@Mod.EventBusSubscriber(modid = ButterfliesMod.MODID, bus = Mod.EventBusSubscriber.Bus.FORGE)
 public class PlayerEventListener {
+
+    /**
+     * Construction
+     * @param forgeEventBus The event bus to register with.
+     */
+    public PlayerEventListener(IEventBus forgeEventBus) {
+        forgeEventBus.register(this);
+        forgeEventBus.addListener(this::onEntityInteractEvent);
+        forgeEventBus.addListener(this::onItemCraftedEvent);
+    }
+
+    /**
+     * Called when a player interacts with an entity.
+     * @param event The interaction event
+     */
+    private void onEntityInteractEvent(PlayerInteractEvent.EntityInteract event) {
+        if (event.getTarget() instanceof Cat cat) {
+            Player player = event.getEntity();
+            InteractionHand hand = event.getHand();
+            ItemStack stack = player.getItemInHand(hand);
+            Item item = stack.getItem();
+            if (item instanceof NameTagItem) {
+                String name = stack.getHoverName().getString();
+
+                //  If the cat is named Snow, then change it to a white cat.
+                if ("Snow".equals(name)) {
+                    CatVariant variant = BuiltInRegistries.CAT_VARIANT.get(CatVariant.WHITE);
+                    if (variant != null) {
+                        cat.setVariant(variant);
+                    }
+                }
+            }
+        }
+    }
 
     /**
      * Handles adding new pages to a butterfly book.
      * @param event The event data
      */
-    @SubscribeEvent
-    public static void onItemCraftedEvent(PlayerEvent.ItemCraftedEvent event) {
+    private void onItemCraftedEvent(PlayerEvent.ItemCraftedEvent event) {
 
         ItemStack craftingItem = event.getCrafting();
         if (craftingItem.getItem() instanceof ButterflyBookItem) {
@@ -81,31 +109,6 @@ public class PlayerEventListener {
                         if (inventoryItem.getItem() instanceof ButterflyBookItem) {
                             ButterflyBookItem.addPage(oldBook, craftingItem, index);
                         }
-                    }
-                }
-            }
-        }
-    }
-
-    /**
-     * Called when a player interacts with an entity.
-     * @param event The interaction event
-     */
-    @SubscribeEvent
-    public static void onEntityInteractEvent(PlayerInteractEvent.EntityInteract event) {
-        if (event.getTarget() instanceof Cat cat) {
-            Player player = event.getEntity();
-            InteractionHand hand = event.getHand();
-            ItemStack stack = player.getItemInHand(hand);
-            Item item = stack.getItem();
-            if (item instanceof NameTagItem) {
-                String name = stack.getHoverName().getString();
-
-                //  If the cat is named Snow, then change it to a white cat.
-                if ("Snow".equals(name)) {
-                    CatVariant variant = BuiltInRegistries.CAT_VARIANT.get(CatVariant.WHITE);
-                    if (variant != null) {
-                        cat.setVariant(variant);
                     }
                 }
             }