+++ src/main/java/com/bokmcdok/butterflies/world/ButterflyData.java
@@ -2,12 +2,11 @@ package com.bokmcdok.butterflies.world;
 
 import com.bokmcdok.butterflies.ButterfliesMod;
 import com.bokmcdok.butterflies.lang.EnumExtensions;
-import com.google.gson.JsonDeserializationContext;
-import com.google.gson.JsonDeserializer;
-import com.google.gson.JsonElement;
-import com.google.gson.JsonObject;
-import com.google.gson.JsonParseException;
+import com.google.gson.*;
+import com.mojang.logging.LogUtils;
 import net.minecraft.resources.ResourceLocation;
+import net.minecraft.server.packs.resources.Resource;
+import net.minecraft.server.packs.resources.ResourceManager;
 import net.minecraft.tags.BlockTags;
 import net.minecraft.util.RandomSource;
 import net.minecraft.world.entity.LivingEntity;
@@ -15,6 +14,8 @@ import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.level.block.state.BlockState;
 import org.jetbrains.annotations.NotNull;
 
+import java.io.BufferedReader;
+import java.io.IOException;
 import java.lang.reflect.Type;
 import java.util.Collection;
 import java.util.HashMap;
@@ -92,7 +93,14 @@ public record ButterflyData(int butterflyIndex,
         FORESTS_AND_PLAINS,
         ICE,
         JUNGLES,
-        PLAINS
+        PLAINS,
+        NETHER,
+        FORESTS_AND_WETLANDS,
+        PLAINS_AND_SAVANNAS,
+        PLAINS_AND_WETLANDS,
+        HILLS_AND_PLATEAUS,
+        FORESTS_PLAINS_WETLANDS,
+        WETLANDS
     }
 
     // Helper enum to determine a butterflies overall lifespan.
@@ -287,7 +295,7 @@ public record ButterflyData(int butterflyIndex,
                         diurnality,
                         extraLandingBlocks,
                         plantEffect,
-                        new ResourceLocation(ButterfliesMod.MODID, breedTarget),
+                        new ResourceLocation(ButterfliesMod.MOD_ID, breedTarget),
                         eggMultiplier
                 );
             }
@@ -311,31 +319,6 @@ public record ButterflyData(int butterflyIndex,
                                                           T fallback) {
             String value = object.get(key).getAsString();
             return EnumExtensions.searchEnum(enumeration, value, fallback);
-        }
-    }
-
-    /**
-     * Create new butterfly data.
-     * @param entry The butterfly data.
-     */
-    public static void addButterfly(ButterflyData entry) {
-        ENTITY_ID_TO_INDEX_MAP.put(entry.entityId, entry.butterflyIndex);
-        BUTTERFLY_ENTRIES.put(entry.butterflyIndex, entry);
-
-        //  Recount the butterflies
-        if (entry.type != ButterflyType.SPECIAL) {
-            int total = 0;
-            for (ButterflyData i : BUTTERFLY_ENTRIES.values()) {
-                if (i.type == entry.type) {
-                    ++total;
-                }
-            }
-
-            if (entry.type == ButterflyType.BUTTERFLY) {
-                NUM_BUTTERFLIES = total;
-            } else if (entry.type == ButterflyType.MOTH) {
-                NUM_MOTHS = total;
-            }
         }
     }
 
@@ -368,12 +351,37 @@ public record ButterflyData(int butterflyIndex,
     }
 
     /**
+     * Create new butterfly data.
+     * @param entry The butterfly data.
+     */
+    public static void addButterfly(ButterflyData entry) {
+        ENTITY_ID_TO_INDEX_MAP.put(entry.entityId, entry.butterflyIndex);
+        BUTTERFLY_ENTRIES.put(entry.butterflyIndex, entry);
+
+        //  Recount the butterflies
+        if (entry.type != ButterflyType.SPECIAL) {
+            int total = 0;
+            for (ButterflyData i : BUTTERFLY_ENTRIES.values()) {
+                if (i.type == entry.type) {
+                    ++total;
+                }
+            }
+
+            if (entry.type == ButterflyType.BUTTERFLY) {
+                NUM_BUTTERFLIES = total;
+            } else if (entry.type == ButterflyType.MOTH) {
+                NUM_MOTHS = total;
+            }
+        }
+    }
+
+    /**
      * Accessor to help get butterfly data when needed.
      * @return A valid butterfly data entry.
      */
     public static ButterflyData getButterflyDataForEntity(LivingEntity entity) {
         String species = getSpeciesString(entity);
-        ResourceLocation location = new ResourceLocation(ButterfliesMod.MODID, species);
+        ResourceLocation location = new ResourceLocation(ButterfliesMod.MOD_ID, species);
         return getEntry(location);
     }
 
@@ -464,12 +472,36 @@ public record ButterflyData(int butterflyIndex,
     }
 
     /**
+     * Load the butterfly data.
+     * @param resourceManager The resource manager to use for loading.
+     */
+    public static void load(ResourceManager resourceManager) {
+        Gson gson = new GsonBuilder().registerTypeAdapter(ButterflyData.class, new ButterflyData.Serializer()).create();
+
+        // Get the butterfly JSON files
+        Map<ResourceLocation, Resource> resourceMap =
+                resourceManager.listResources("butterfly_data", (x) -> x.getPath().endsWith(".json"));
+
+        // Parse each one and generate the data.
+        for (ResourceLocation location : resourceMap.keySet()) {
+            try {
+                Resource resource = resourceMap.get(location);
+                BufferedReader reader = resource.openAsReader();
+                ButterflyData butterflyData = gson.fromJson(reader, ButterflyData.class);
+                ButterflyData.addButterfly(butterflyData);
+            } catch (IOException e) {
+                LogUtils.getLogger().error("Failed to load butterfly data", e);
+            }
+        }
+    }
+
+    /**
      * Gets the resource location for the caterpillar item.
      * @return The resource location of the caterpillar item.
      */
     public ResourceLocation getCaterpillarItem() {
         if (this.entityId != null) {
-            return new ResourceLocation(ButterfliesMod.MODID, "caterpillar_" + this.entityId);
+            return new ResourceLocation(ButterfliesMod.MOD_ID, "caterpillar_" + this.entityId);
         }
 
         return null;
@@ -481,7 +513,7 @@ public record ButterflyData(int butterflyIndex,
      */
     public ResourceLocation getButterflyEggItem() {
         if (this.entityId != null) {
-            return new ResourceLocation(ButterfliesMod.MODID, entityId + "_egg");
+            return new ResourceLocation(ButterfliesMod.MOD_ID, entityId + "_egg");
         }
 
         return null;
@@ -492,7 +524,7 @@ public record ButterflyData(int butterflyIndex,
      * @return The resource location of the butterfly.
      */
     public ResourceLocation getButterflyEntity() {
-        return new ResourceLocation(ButterfliesMod.MODID, this.entityId);
+        return new ResourceLocation(ButterfliesMod.MOD_ID, this.entityId);
     }
 
     /**
@@ -500,7 +532,7 @@ public record ButterflyData(int butterflyIndex,
      * @return The resource location of the butterfly egg.
      */
     public ResourceLocation getButterflyEggEntity() {
-        return new ResourceLocation(ButterfliesMod.MODID, this.entityId + "_egg");
+        return new ResourceLocation(ButterfliesMod.MOD_ID, this.entityId + "_egg");
     }
 
     /**
@@ -508,7 +540,7 @@ public record ButterflyData(int butterflyIndex,
      * @return The resource location of the caterpillar.
      */
     public ResourceLocation getCaterpillarEntity() {
-        return new ResourceLocation(ButterfliesMod.MODID, this.entityId + "_caterpillar");
+        return new ResourceLocation(ButterfliesMod.MOD_ID, this.entityId + "_caterpillar");
     }
 
     /**
@@ -516,7 +548,7 @@ public record ButterflyData(int butterflyIndex,
      * @return The resource location of the chrysalis.
      */
     public  ResourceLocation getChrysalisEntity() {
-        return new ResourceLocation(ButterfliesMod.MODID, this.entityId + "_chrysalis");
+        return new ResourceLocation(ButterfliesMod.MOD_ID, this.entityId + "_chrysalis");
     }
 
     /**